♻️  Refactor: rename to useGridEventMouseDown and cleanup
