+++ packages/web/src/views/Calendar/components/Grid/AllDayRow/AllDayEvents.tsx
@@ -7,7 +7,7 @@ import { selectDraftId } from "@web/ducks/events/selectors/draft.selectors";
 import { selectAllDayEvents } from "@web/ducks/events/selectors/event.selectors";
 import { draftSlice } from "@web/ducks/events/slices/draft.slice";
 import { useAppDispatch, useAppSelector } from "@web/store/store.hooks";
-import { useGridEventMouseHold } from "@web/views/Calendar/hooks/grid/useGridEventMouseHold";
+import { useGridEventMouseDown } from "@web/views/Calendar/hooks/grid/useGridEventMouseDown";
 import { Measurements_Grid } from "@web/views/Calendar/hooks/grid/useGridLayout";
 import { WeekProps } from "@web/views/Calendar/hooks/useWeek";
 import { AllDayEventMemo } from "./AllDayEvent";
@@ -46,7 +46,7 @@ export const AllDayEvents = ({
     );
   };
 
-  const { onMouseDown } = useGridEventMouseHold(
+  const { onMouseDown } = useGridEventMouseDown(
     Categories_Event.ALLDAY,
     handleClick,
     handleDrag,
